# Debug: 1. cache refresh at every init; 2. compute cardinality only once