@MarkDana 2021-12-26
Closed this pr - the two issues (cache for different data & pass cardinality to chisq/gsq) has already been covered in @chenweiDelight 's newest commit (732b9d1).